[connectivity] Move away from invaid use of cpp20::span for C++20
Our current cpp20::span polyfill accepts types with data() and size()
methods as ranges for its range constructor, but this is not the case
for the real std::span which cpp20::span falls back to when available.
Instead we now use the constructor that does take a pointer and size.

Test: Add --experimental_cxx_version=20 to args.gn and fx build
Change-Id: I407cc51ac6d89bbebf2bde865b90e3958a1050ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632499
Reviewed-by: John Sheu <sheu@google.com>
Commit-Queue: Mark Schott <schottm@google.com>