[run-test-suite] Only output prefix when needed
In some infra builders the prefix is spammed for each line, which makes
the output very difficult to read. To resolve this issue, we keep track
of the last handle that wrote to the output stream, and only output the
prefix when the handle has changed.

Testing: fx test run-test-suite-lib-tests

Change-Id: I33640691ef94372ae0fc4e693e70b579f4ad905a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621282
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>