[sshutil] Mark ErrClosed as a connection error
This can happen when a keep-alive ping times out and the connection gets
closed in another goroutine.

This will cause testrunner to not consider a test to have failed if it
gets an ErrClosed, and to instead reconnect and retry running the test.

Bug: 91584
Change-Id: I2aa286d3e36c2bc5db41f16e3596fa44f6b9e051
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633302
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>