[tools][fuzz] Switch undercoat to C++ symbolizer
The Go symbolizer was removed from the base build in fxrev.dev/629945.
This CL switches over to using the C++ symbolizer, which also no longer
has a dependency on `llvm-symbolizer`.

Additionally, failing to be able to run the symbolizer was causing
fuzzer runs to hang, so this CL also improves handling of fuzzer output
processing to exit immediately with better errors messages in this sort
of circumstance.

Lastly, a new QEMU parameter was added to stop SeaBIOS from emitting
characters that interfered with terminal output and that was making it
slightly harder to read logs.

Bug: 91640
Change-Id: I58f67c340d85ba0dc5dbbef38a114aebe79f2aa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633494
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>