[netstack3][ndp] Refactor IPv6 device addr config
Refactor the values in AddressEntry to constrain the presence of the
valid_until field to only those entries corresponding to
SLAAC-configured addresses. This makes it impossible to create an entry
for a manually-configured address with a non-empty valid_until value. It
also prepares the way for us to add additional information for
SLAAC-specified addresses in the future.

Bug: 21428
Change-Id: Iaeaf30b57650f1efae1afa377dc14037a219b343
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/623032
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>