[component_manager] Replace `AbsoluteMoniker` constructor
Instead of creating an `AbsoluteMoniker`, then turning it into a
`PartialAbsoluteMoniker`, this change creates a `PartialAbsoluteMoniker`
directly, when used to obtain the instance id from a component. To do
so, `.to_partial()` is called on the existent `RelativeMoniker`
(required by `storage`) in `PartialAbsoluteMoniker::from_relative()`.

No behavioral change expected. This is refactoring only.

The reason for this change is to reduce the occurrences of
`AbsoluteMoniker`s if they are not needed, even in component manager.

Bug: 87407
Change-Id: If6dc557d5ca33c82d552c3f8a7c48eb36c58a5aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632998
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>