[wlan][sae] Parameterize HMAC constructions
Thus far the kdf_sha256 implementation and the HMAC construction used in
our SAE support are hard-coded to use a SHA256 HMAC.  This is correct
for now, as we only support IKE Group 19 (IETF RFC 2409) which requires
SHA256.

This commit refactors the HMAC constructions to be parameterizable based
on the type of HMAC, which make these constructions more widely useful
and simplifies future support for other groups.

This commit also renames usages of the "h" function to "hkdf_extract",
as this function is precisely the function described in IETF RFC 5869,
and adds an implementation of the corresponding "hkdf_expand", which
will be used for WPA3 Hash-to-Element support.

Test: fx test wlan-sae-tests

Change-Id: Ifc5067728310c8b9e539314c3971cee1689732e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632094
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Sheu <sheu@google.com>