[fidl][cpp] No more inheriting from ClientBase
At this point, `WireClientImpl` only has async methods, so we can
replace it with `WireWeakAsyncClientImpl` and no longer inherit from
`ClientBase`. The more scalable pattern is to borrow a `ClientBase*`
and call methods on that.

Some tests were written with inheritance in mind. Those are also changed
to taking a `ClientBase*`.

It would be better to rename |ClientBase| to |ClientMessenger| or
similar since it is no longer a base class, but doing that later leads
to less intense churn.

Fixed: 82085

Change-Id: I635318321d2c42b6e169543159df5cc33a95ac76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632682
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>