[ffx] Add --persist to ffx emu shutdown
One request we've received for the emulator is to have a mechanism for
persistent state for debugging, reuse, etc. This CL adds a flag to the
shutdown subcommand, "--persist", which instructs the tool not to remove
the state directory on shutdown.

This means after an instance is shut down, it will still show up in the
`ffx emu list` results. To help with this, we add a status indicator to
those results.

This also means we need to be pickier about the entries we return in the
list subcommand: the results should only include directory entries, and
they should contain an engine.json file; otherwise, they can't be valid
instances.

Change-Id: Id6afcd0ebe8e6b3065037f439080b268986a086f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634102
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>