[uart] Fix equality operators for C++20
uart-mock relies on comparing cpp17::variants of custom structs with
their own operator==, and the way comparison operators work in C++20
changed enough to break and make it look like mock-function was the
problem. Changing one of the operators to a friend removes all ambiguity
and works in both C++17 and C++20.

Test: uart-unittests

Change-Id: I9d884a7356d5bbc9e11720fbe360ab86acb153a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631725
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Mark Schott <schottm@google.com>