[paver] Fix std::span arguments for C++20
Our cpp20::span polyfill falls short, but the real std::span in C++20
doesn't work on temporaries to prevent lifetime issues. In this case the
data is copied out anyway so it's fine, but we still need to use lvalue
for span parameters for it to compile.

Test: Add --experimental_cxx_version=20 to args.gn and fx build
Change-Id: I5b3925046e4add8d3ba982554b5f18903ebc10dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634212
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Mark Schott <schottm@google.com>