[test_manager] Move facet retrieval to a new mod
This moves facet retrieval and parsing to another module. This will be
useful as for debug data, we need to inspect the facets, then notify
the debug_data component of the moniker of the test suite (which
depends on the collection), then launch the suite.

Bug: 77634
Testing: fx test test-manager-unittests

Change-Id: I9cdb1c10462b767540119d0e35e43b037e7897db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631763
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>