[security][swd] Introduce common configuration for securty pkg tests
Much of the test configuration needed for security pkg tests
is not specific to a particular product. This change introduces
common configuration in a way that is sufficiently abstract to add
product-specific configuration as well. The code includes
placeholder component configuration data so that tests can be run
in-tree without additional files or template instantiations.

Bug: 91099
Test: fx test fuchsia-pkg://fuchsia.com/core_access_ota_blob_as_executable#meta/core_access_ota_blob_as_executable.cm

Change-Id: I03a9b580df27854c64d7c57c9aea22a93863a358
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629501
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>