[fidl] Remove Struct.IsRequestOrResponse from fidlgen
To support named payloads, structs will no longer be allowed to "know"
whether or not they are used as a transactional message body, as it will
now be possible for a single struct to be used in both contexts.  This
change is the first in a series that will remove the
"is_request_or_reponse" JSON IR field, starting off by removing all
usages of the field in fidlgen.

Test: fidldev test all
Bug: 88344
Change-Id: Ie13c7d9262afb9d78a4c2863256e59c493b9f433
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632581
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>