[scripts][fuzzing] Switch to cpp_symbolizer
The legacy fuzzing scripts were using the legacy Go symbolizer. Until
they can be deprecated, they need to switch to the newer C++ symbolizer.

Note that this breaks the coverage workflow due to a dependency on the
`-json-output` parameter of the old symbolizer. fxbug.dev/91703 tracks a
fix for that issue.

Bug: 91640
Change-Id: Ibbb90d16874dbee62b44325fd91966e01328222c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633223
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>