[zxio] Let caller provide storage for dirents when enumerating
zxio provides a facility for enumerating over directory entries and
placing data into a zxio_dirent_t. This change lets the caller provide
the storage for the entries rather than returning a pointer to a zxio
internal pointer. This removes one data copy for the entries and
simplifies some of the lifecycle reasoning. The name field is typically
intended to go into a dirent struct, so this also lets the caller
specify where the name bytes should go in the zxio_dirent->name pointer.

With this change zxio is still buffering entries returned by the
filesystem server and not yet consumed by the client in an internal
buffer. A follow-up will let the caller also supply storage for this
buffer to remove the heap dependency inside zxio.

This API is used by https://fuchsia.googlesource.com/third_party/mesa
which is updated independently from fuchsia.git. To facilitate the API
transition this change adds a preprocessor define to zxio's public
configuration which mesa can use to select either the old or new API.
Once mesa is updated, we can remove this configuration.

Bug: 91172
Change-Id: Ic7eeea8506e258d35b183a12cd04b5d594d2c89a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628538
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: James Robinson <jamesr@google.com>