[component_manager] Hub returns error for instance not found
This CL replaces direct array accesses and `expect()`s for instances in
the `instance_map` with returning an error. If an instance is not found
in the `instance_map`, then a `ModelError` is returned, instead of a
panic. This emphasizes the responsibility of the caller to appropriately
handle `ModelErrors` produced from the hub event hooks.

The reason for this CL is prevent a panic and provide the opportunity
to handle a failure scenario, when an instance is not found in hub.rs.

Multiply: component-manager-tests: 1000
Multiply: hub_integration_test: 1000
Test: fx test component-manager-tests
      fx test hub_integration_test
Change-Id: I798fa1fdbaa669030aa29a4521d9994606be181f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633767
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>