[fxfs] Move serialization into serialized_types.
This is an incremental change towards versioning but doesn't implement
any versioning changes yet. This change is intended as both an
illustration and a proof-of-concept but I expect further work to be
required before it's suitable to use.

This replaces calls to bincode serialize_into/deserialize_from with
equivalent calls to [Version] trait methods. The LSM tree code will be
tackled in a separate CL due to complications in the way it's
implemented (e.g. different types with the same layout are
serialized/deserialized).

Change-Id: I99f570dd70aba848c1ff96f7c4edeefb370d61c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622842
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>