[kernel][vm] Remove page_request_t from PageRequest
The only fields of page_request_t that are used PageProvider
implementations are the offset, len, type and queue_node. The first
three of these are replicated from PageRequest, and the queue_node can
be added to PageRequest. The rest of the page_request_t fields are not
used and so the entire page_request_t can be dropped out of
PageRequest.

Change-Id: Idd418388968a29416f174e1fb4e454b46afabfd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631344
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>