[fdio][zxio] Use buffer+length strings in two path operations
This updates the logic for operations performed on two paths (rename and
link) to represent the string using an explicitly tracked length instead
of relying on null termination in both the zxio and fdio layers.

In zxio, the C interface is updated to accept a pointer to a buffer and
a length and the implementation is updated to use this length instead of
calling strlen() on the passed in value.

In fdio, the buffers used for these operations are allocated internally
on the stack. This updates the logic for allocating and manipulating the
path buffers to use fbl::StringBuffer which keeps track of the length
and uses std::string_view to pass buffer+length tuples inside fdio.

fbl::StringBuffer is also allocating and managing a null terminator, as
is the old code. This terminator isn't necessary and we may wish to
remove it later on.

Bug: 91672
Change-Id: I0708c21019df6423af1bce3f29146312562c6423
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634644
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>