[system-update-checker] Grant fuchsia.space.Manager
When system-update-checker can run out of space when trying to update the
tuf metadata and fetch the update package to see if it should be installed.
Unfortunately when we transitioned over to CFv2, we forgot to include
passing fuchsia.space.Manager. We didn't catch this, because we didn't have
an integration test to make sure that we actually call Gc when we run out
of space.

This fixes the routing, and adds an integration test to make sure
everything works correctly.

Fixed: 91884
Change-Id: I610f391a382f3747d3cd92d8cf7630816f072f96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635370
Reviewed-by: Ben Keller <galbanum@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>