[wlan][iwlwifi] Move iwl_mvm_sta ownership external to driver core
In the Linux driver, the iwl_mvm_sta instance is allocated as the
"driver-private" part of the ieee80211_sta struct, and hence is not
owned by the driver core.  We do the same here by introducing a MvmSta
class which maintains ownership of the iwl_mvm_sta, and which is held by
the WlanSoftmacDevice at a layer above the iwlwifi driver core.

Bug: 75293, 86715, 86728, 90919
Test: fx test iwlwifi_test, booted and connected to WLAN on NUC

Change-Id: I69e1b3c3e299a24b62bb40ad19b0663c7abd6875
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626387
Reviewed-by: Yung-chieh Lo <yjlou@google.com>
Commit-Queue: John Sheu <sheu@google.com>