[debug_data] Ensure handle used for Wait is valid on Cancel
Move the inner state of the DataProcessor class to the data processor
thread so that the Zircon event handle it contains is still valid when
the Wait is cancelled. We need to do this as the underlying async
libraries assume the handles will be valid at the time a wait is
cancelled:

[00035.034491][1187][1312][processor] WARNING: ASSERT FAILED at (../../zircon/system/ulib/async-loop/loop.c:551): status == ZX_ERR_NOT_FOUND
[00035.034587][1187][1312][processor] WARNING: zx_port_cancel: status=-11

Testing: This issue only arises once we switch to using the new debug
data flow, in which the affected destructor is called. Verify by
running CQ with the change to use the new debug data flow.

Change-Id: Ifab8f3b23db071366e5fa321e6dba8169cf17945
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633765
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>