[fidl] Enforce visibility for CPP bindings.
(Thanks to Aaron Wood for doing the hard work to isolate this).

We were not currently enforcing visibility constraints on the CPP
bindings for a FIDL build rule, unlike other languages. Since we tested
on other languages we didn't actually realize this.

Fortunately it looks like we only picked up two extra dependencies, one
in forensics which looks list just on omission in the visibility list,
and one in stash which I've created a bug to resolve.

Change-Id: I83beb79d2e1b486b4cfd299fa827db4612cebb1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632582
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>