[physboot][handoff] Don't forget to report boot.timeline.phys-handoff
Fix a small bug in the predicate of a for loop which was causing the
kCount member of an enumeration to get skipped.  The net result of
this was that the `boot.timeline.phys-handoff` kcounter was always
staying with a value of 0, which was causing the
kernel-boot-benchmarks test/metric-collector to ignore the value
entirely.

Change-Id: I8e858f9620e2399ff6ff951d5493d210914c807b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634741
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>