[scrutiny] Use `PartialAbsoluteMoniker` in error/warning message
This CL prints the instance's `PartialAbsoluteMoniker` instead of
`AbsoluteMoniker` on error in `check_instance`; the instance ids of the
instance will not be included when its absolute moniker path is printed.

The reason for this change is to remove `AbsoluteMoniker` references
from public APIs.

Bug: 87407
Change-Id: I76a99d1d79035cbebb51ab5bb6d1aec78c213b01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633581
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>