[zxio] Provide scratch buffer storage inline in zxio_dirent_iterator_t
The zxio directory iteration interface provides entries one at a time
for ease of use. The underlying transport operates in batches of entries
for efficiency, so something has to buffer the results until the caller
is ready to consume them. This change allocates storage for this
buffering inside the zxio_dirent_iterator_t structure.

Since this increases the size of the iterator considerably (from 64
bytes to slightly over 64 KiB) this also updates the fdio usage in
unistd.cc to allocate its iterator in the heap. Tests still generally
allocate this type on the stack. The other direct use of this interface
in the Mesa library is already allocating zxio_dirent_iterator_t in the
heap, so it is unchanged.

This also moves the type definition from zxio/zxio.h to zxio/types.h to
be more consistent with other type definitions.

Bug: 91172
Change-Id: Ia8f677f57122dff7f099af063dc7b9a07706b3a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629583
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>