[pkg-resolver] BlobFetcher type alias -> wrapper struct
This no-op change is just to make it a bit easier to jump to the
fetching logic from the place where the fetch is enqueued. Hopefully it
makes pkg-resolver a little easier to navigate.

A pleasant side-effect is the doc comments get a bit nicer to read,
since they're a lot less abstract (since the queue's processing
behavior is known rather than being arbitrary)

This is a follow up to https://fxrev.dev/628587, which does the same for
the package fetch queue.

Change-Id: I403dcca1d12195fd840c4e61874c8087589eb4f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630920
Fuchsia-Auto-Submit: Dan Johnson <computerdruid@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>