[run_test_suite] Custom run-scoped directory artifacts.
This CL supports retrieving a custom directory as an artifact from the
TestRun. The custom directory is scoped to the run itself rather than
any individual test case. This will be used in the future to support
including information about the complete state of a run to host-side
tools.

Change-Id: I7ca239f901fed2d42d310df2eac20e8be8ea05ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619326
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>