[run-test-suite] Port early termination options to run-test-suite
Copies over some early termination options in ffx test to
run-test-suite. Since the tools use the same underlying library, this
simply exposes some options that are already available to the command
line.
This will simplify a fix to fx test that relies on these options. Since
fx test currently supports a fallback to run-test-suite, this is
necessary to ensure the options behave consistently when the fallback is
used and when it is not used.

Testing: Manually test behavior with `fx shell run-test-suite`
Bug: 88478

Change-Id: I5dd8bf7f89ab8db5f3c963f3e880a9c261b6bb4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635375
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>