[brcmfmac] Handle suspend to prevent reboot crash
The driver would crash on every reboot because threads kept running
during shutdown. This happened because the driver was never notified of
the shutdown and so could not stop threads in an orderly fashion.

Fix this by making the Device class Suspendable and shutting everything
down when Suspend is called.

Bug: 90694
Test: Manual rebooting and verifying no crashing
      fx test sim_device_test brcmfmac_sim_test brcmfmac_test
Change-Id: Id16c64a8b167c8cdc221be8320a83eafa68fb7f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635163
Reviewed-by: John Sheu <sheu@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>