[ffx] Manually add emu in user mode to ffx target
Before the rewrite, ffx emu would make a call to `ffx target add` to add
the emulator instance to the target list when the emulator was running
in user networking mode (i.e. SLiRP). This was needed because the IP
address included in any mDNS advertisements when running with user
networking would contain the internal IP address, which is invalid
outside of the emulation.

This CL reinstates that functionality in the new emu tool, adjusting for
changes to the `ffx target` code in the interim.

This introduces a bit of complexity when the emulator is running in
console/monitor mode, since the process is responsible for initiating
cleanup, but the shutdown code is also async. The "self" parameter from
the start code can't be shared between the threads launched for the two
commands, so we have to split out the parameters we want and pass them
explicitly and make the qemu_base::shutdown_emulator method an
associated function instead.

Bug: 91642
Change-Id: I03cf99c3d5894e4d4fca251d72f98e19018d6545
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634807
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>