[build] extend fastboot_flash image property
This CL documents the already-used fastboot_flash image property, which
is a list of partition names the image should be flashed to in a
standard flash operation.

In addition, this adds a special case that images can supply an empty
list in this property to indicate flash-adjacent images which should be
provided when flashing but aren't flashed directly to a partition, for
example unlock credentials and images used for bootstrapping.

Test: `fx flash` and `flash.sh` still work locally (Sherlock)
Test: `fx flash-remote` pulls unlock creds and bootstrap images
Change-Id: I8fe9500ad434b62c301f191e0163ec064abee82b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/619061
Reviewed-by: Matt Boetger <boetger@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Amit Uttamchandani <amituttam@google.com>