[net-declare] Rename CidrAddr and add tests
The CidrAddr type is misnamed; it represents an IP and prefix, not a
valid CIDR range. Rename it in preparation for adding another type that
represents a valid CIDR range. Add tests of the existing functionality
(with the new name) since there weren't any.

Change-Id: Id259f7014164ea621651ab9ae19f30a6b971d5a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635483
Reviewed-by: Tony Gong <gongt@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>