[llcpp] Assert that server is channel when writing epitaphs
Reason::kClose should only happen when an epitaph is supposed to be
written, which is only supported on channels. This CL adds an assertion
to ensure that the server transport is a channel when a teardown with
Reason::kClose is taking place.

Change-Id: I0431b39de071ebc56143f2adfd6c6a21501de412
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628181
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>