[tcpdump] Replace netdump with tcpdump for fx sniff
Replace the view option with a single wireshark option which can be used
to view a packet capture on wireshark in real-time. Note that this drops
the hex view but we do not seem to have any clients using that feature.
Note that the default view is supported by not including any option.

Bug: 91943

Test:
- fx sniff ethp0003
- fx sniff -t 10 ethp0003
- fx sniff --file test ethp0003
- fx sniff --wireshark ethp0003
- fx sniff --wireshark --file test ethp0003
- fx sniff --wireshark --file test -t 10 ethp0003

Change-Id: I200f4329c70e4b70a26f9b324776487598405f1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635603
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>