[kernel][vm] Reset request on PagerProxy::ClearAsyncRequest
The PageSource calls PagerProxy::ClearAsyncRequest when it wants to
cancel a page request, either because the request has been fulfilled or
aborted by the VMO. It is invalid for PagerProxy to continue holding a
reference to the page request after ClearAsyncRequest, as the ownership
of the request has been transferred back from the PagerProxy to the
PageSource.

Set active_request_ to nullptr on a call to ClearAsyncRequest. This
prevents the PagerProxy from holding a stale pointer to the page
request. The PagerProxy is free to release the corresponding port packet
later on PagerProxy::Free if the ClearAsyncRequest races with a port
dequeue by the user pager. The logic for queueing of new packets remains
unchanged as it is still controlled by the port packet being busy, not
the active request.

Bug: 91935
Change-Id: I58c2c2eee9d534c04183b7f735ab6dba2c415159
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636181
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>