[assembly_util] remove lifetimes from traits and implement on &mut T
I think this is a bit simpler because you don't have to thread the
lifetimes around everywhere and I think it makes more conceptual sense
to have the trait not be generic over the lifetime and leave the
lifetime with the type.

As far as I can tell this doesn't matter to callers because method
dispatch still finds this trait no problem (I didn't have to change the
tests).

Change-Id: I8b03e56d856c627bb7b84ccd3679fcfd3e880668
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635824
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Dan Johnson <computerdruid@google.com>