[ffx] Clean up error messages for deserialization
As we've been developing ffx emu, there hae been a lot of changes to the
serialized configuration along the way. Each time a state file is left
on the disk and the configs change, we hit errors on start-up and
shutdown about some field missing, but no indication of how to fix it.

This CL adds text for these serialization errors to indicate 1) where
they came from, and 2) what to do about them. We modify the output for
both the start, shutdown, and show subcommands.

Change-Id: Iaafdcd17fb194174e6b7584d7853d6b5b02d10c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635982
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>