[ffx] Add graph orientation flag.
This lets the user change the orientation of the graph to left-to-right via `ffx component graph --orientation=left_to_right`.

This change also changes the default orientation from left-to-right
to Graphviz's default of top-to-bottom.
This tends to look better although it takes up more room.

Change-Id: I33f8aa952dd230611600e2c553f1635800e81862
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632601
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>