[fshost] start fshost rewrite
This adds an experimental rust version of fshost. Currently the rust
version implements fuchsia.fshost.Admin for shutting down, and that's
it.

An integration test harness is added, using RealmBuilder, which tests
behaviors against both the rust and the cpp version of fshost to build
confidence in the replication of these behaviors in the rewrite.

Bug: 87511
Test: fx test fshost_integration_tests
Change-Id: I475e399ecc9ca3e306f017d6cd4f11842a56aed9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598882
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>