[tools][fuzz] Disable kernel lockup detector in fuzzing VMs
Running in GCE frequently causes kernel oops/panics when the host
doesn't give the VM enough time to run. Since this is expected behavior,
this CL tells Zircon to ignore this condition.

Additionally, the SSH reconnect behavior was made even more generous to
account for this possibility, and the kernel log was set to be
unbuffered to ensure we aren't missing any final messages.

Bug: 92109
Bug: 41327
Change-Id: Ia7207add1681fd63ff555d66911115251aa5c2cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637008
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>