[component_manager] Hub permits absent parent at child discovery
The CL returns Ok, instead of an error, if a parent component is not
present in the `instance_map` at the time its child is discovered.

Additionally, the `Hub::add_instance_if_necessary` call was refactored
to return early if `None` is returned (ie the target component is
already in the instance map) to reduce drift with nested conditionals.

The reason for the CL is to address the scenario where a purge event for
the parent has removed it from the instance map before a discovered
event for its child has been processed. A TODO is added for
fxbug.dev/89503 to investigate this event ordering scenario further.

Bug: 89503
Change-Id: I59f1da28169d4385007610f63a40228bf9fd42a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635825
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>