[fx][rbe] Inform users if RBE tools are missing
After an `fx args` or `fx set`, if RBE is enabled,
make sure needed tools like `gcloud` are available,
or inform the user how to install them.

This moves this check earlier in the developer workflow, rather than
catching it at `fx build` time in the 'fuchsia-reproxy-wrap.sh` stage.

re-client tools are already under prebuilt, so no need to check for
them.

Even better would be to eliminate the use of gcloud altogether... in the
future.

Tested:
* ran `fx set --rbe`
* ran `fx args`; *edit* "enable_rbe = true"
* repeated above without gcloud

Bug: 91579
Change-Id: I9929ec25ec267a6975a67f7d6a5b6da8824e9402
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635364
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Fang <fangism@google.com>