[fxfs] Eagerly reserve metadata space
The global metadata reservation should always be created before any
transactions are attempted, and it is an invariant that we should always
have enough space for that reservation.

Because this reservation is lazily initialized, it is possible for it to
be initialized after some other transaction has stolen space from the
allocator, e.g. if the first transaction requires its own reservation.

Fix this to eagerly reserve the space when we start up.

Also add tests for low space conditions to fs_test, since that is what
triggers this bug.

Fixed: 91475
Change-Id: I82133b470bd6e522bcfbc32a536a69fd78180274
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631772
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>